+++ albomon/core/src/Infrastructure/UI/Cli/Command/CheckCustomCatalogComand.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Albomon\Core\Infrastructure\UI\Cli\Command;
 
 use Albomon\Core\Application\MonitorApplicationService\MonitorApplicationService;
+use Albomon\Core\Infrastructure\UI\Cli\Traits\SymfonyStyleTrait;
 use DateTime;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Helper\Table;
@@ -14,6 +15,8 @@ use Symfony\Component\Console\Style\SymfonyStyle;
 
 class CheckCustomCatalogComand extends Command
 {
+    use SymfonyStyleTrait;
+
     private const CATALOG_FILE_NAME = 'custom-catalog.json';
 
     /** @var MonitorApplicationService */
@@ -36,17 +39,19 @@ class CheckCustomCatalogComand extends Command
     protected function execute(InputInterface $input, OutputInterface $output): ?int
     {
         $alboList = $this->getCustomCatalog();
-        // TODO remove
+
         $AlboPopSpecValidation = 'Non Rilevato';
 
-        $io = new SymfonyStyle($input, $output);
-
         if ($input->isInteractive()) {
+            $io = $this->getSymfonyStyle($input, $output);
+
             $io->text('Inizio scansione albi, origine dati: '.self::CATALOG_FILE_NAME);
 
             $io->text('Il catalogo albi contiene '.count($alboList).' feed da analizzare.');
 
-            $io->note('Il tempo necessario alla scansione può variare in base al tipo di connessione e  alle condizioni della rete.');
+            $io->note('Il tempo necessario alla scansione può variare in base al tipo di connessione ed alle condizioni della rete.');
+
+            $io = new SymfonyStyle($input, $output);
         }
 
         $monitorResultCollection = $this->monitorService->checkAlboList($alboList);
@@ -66,7 +71,6 @@ class CheckCustomCatalogComand extends Command
                 if (!$monitorResult->httpStatus()) {
                     $table->appendRow([$monitorResult->feedUrl(), sprintf('<error>%s</error>', 'NON ATTIVO'), $AlboPopSpecValidation, '', 'server error']);
                 } else {
-                    // TODO format output per data aggiornamento visualizzare colore diverso per ritardo maggiore di 1 settimana  1 mese 1 anno
                     $lastFeedItemDateWithDifference = $this->formatContentUpdatedAt($monitorResult->lastFeedItemDate());
 
                     $table->appendRow([$monitorResult->feedUrl(), 'ATTIVO', $AlboPopSpecValidation, $lastFeedItemDateWithDifference, '']);
@@ -86,7 +90,7 @@ class CheckCustomCatalogComand extends Command
         if (!file_exists($catalogFile)) {
             throw new \RuntimeException('Catalog file not found. file: '.$catalogFile);
         }
-        // TODO check exist before open throw exception or message in console
+
         $strJsonFileContents = file_get_contents($this->catalogDir.DIRECTORY_SEPARATOR.self::CATALOG_FILE_NAME);
 
         $customCatalog = json_decode((string) $strJsonFileContents, true);